metadata-extraction-service/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
mysql-connector-python==8.0.33
pillow==9.1.1
pika==1.2.1
requests==2.28.2